rename DreamsListItem to plural
